Fix profile number handling in case HID_CMD_SETPROFILE in hidGetOutReport()
The old version got the profile number from the first hid argument. This prohibited setting a profile other than profile 0, because the first hid argument is used to send the start address of the data.

When the full profile (params + info) is received, the data is first checked. The old profile is saved if necessary. The params of the new profile are written to the dataflash, overwriting the active profile. It is then saved again to the slot with the profile number defined in params.Profile, overwriting the old profile in this slot.

This should resemble the original intended function.